Added MDB, new LDB methods, SERV.ITEMDEF/CHARDEF.ISTEVENT. - Added: MDB object. It's an SQLite database instance working like LDB, with the fundamental difference that MDB works with a temporary, in-memory database (this means that all the data is lost when you close the server). Use it if you have small databases where you query from very frequently, since accessing a table from a db file is much slower than doing that from an in-memory db. OPEN and CLOSE methods doesn't work with MDB. - Added: LDB/MDB functions IMPORTDB and EXPORTDB "database.db". The first loads into the database object a database file, the second saves the database object to a database file. - Added: LDB.FILENAME property to get the filename of the database connected to. Returns an empty string if there isn't an open connection. - Fixed: Crash when trying to execute a QUERY on a closed/never opened LDB. - Added: ISTEVENT for SERV.CHARDEF and SERV.ITEMDEF.

+++ src/network/network_multithreaded.cpp
@@ -1940,7 +1940,7 @@ NetState* NetworkThreadStateIterator::next(void)
 	{
 		// current thread, we can use the thread's state list directly
 		// find next non-null state
-        const size_t sz = m_thread->m_states.size();
+        const int sz = int(m_thread->m_states.size());
 		while (m_nextIndex < sz)
 		{
 			NetState* state = m_thread->m_states[m_nextIndex];
